requirements.txt
python-telegram-bot==13.1
boto3==1.17.3